LayoutTests/fast/dom/Geolocation/enabled-expected.txt

Tests that the navigator.geolocation object is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof navigator.geolocation == 'object' is true
PASS hasGeolocationProperty() is true
PASS 'geolocation' in navigator is true
PASS navigator.hasOwnProperty('geolocation') is true
PASS successfullyParsed is true

TEST COMPLETE
